refactor: Split test cases in many test files LuxImportServiceTests ManifestRepositoryTests
